# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "hoydigo API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/political-party?page=1": {
            "get": {
                "tags": [
                    "Admin - Political Party"
                ],
                "summary": "List of the political parties",
                "description": "List of the political parties",
                "operationId": "listPoliticalParties",
                "responses": {
                    "200": {
                        "description": "List of political parties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                                        "type": "string",
                                                        "format": "text",
                                                        "example": "CECOL"
                                                    },
                                                    "name": {
                                                        "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "Centro Colombia"
                                                    },
                                                    "description": {
                                                        "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                                        "type": "string",
                                                        "format": "text",
                                                        "example": "Descripcion de centro colombia"
                                                    },
                                                    "political_position": {
                                                        "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                                                    "type": "string",
                                                                    "example": "CEN"
                                                                },
                                                                "name": {
                                                                    "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                                                    "type": "string",
                                                                    "example": "Centro"
                                                                },
                                                                "description": {
                                                                    "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                                                    "type": "string",
                                                                    "example": "Se conoce por centro ..."
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "example": {
                                                            "id": "CEN",
                                                            "name": "Centro",
                                                            "description": "Se conoce por centro ..."
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": "MIPA2",
                                                    "name": "Mi partido",
                                                    "description": "Description del partido",
                                                    "political_position": {
                                                        "id": "CEN",
                                                        "name": "Centro",
                                                        "description": "Se conoce por centro ..."
                                                    }
                                                },
                                                {
                                                    "id": "MIPAR",
                                                    "name": "Mi partido",
                                                    "description": "Description del partido",
                                                    "political_position": {
                                                        "id": "CEN",
                                                        "name": "Centro",
                                                        "description": "Se conoce por centro ..."
                                                    }
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                            "type": "string",
                                            "format": "text",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User without permission to thi endpoint",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Returns all political parties\n\nIt is returning the political parties in pagination\nthe amount of parties per page is 20.",
                                            "type": "string",
                                            "format": "text",
                                            "example": "Permission denied."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "name": "Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}